requirements.txt
django>=5.0,<6.0
djangorestframework>=3.14,<3.15
psycopg2-binary>=2.9,<3.0
djangorestframework-simplejwt>=5.4,<6.0
drf-spectacular>=0.27,<0.28
